+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/BotCommand.kt
@@ -9,4 +9,13 @@ data class BotCommand(
     val command: String,
     @SerialName(descriptionField)
     val description: String
-)
+) {
+    init {
+        if (command.length !in botCommandLengthLimit) {
+            error("Command size must be in range $botCommandLengthLimit, but actually have length ${command.length}")
+        }
+        if (description.length !in botCommandDescriptionLimit) {
+            error("Command description size must be in range $botCommandDescriptionLimit, but actually have length ${description.length}")
+        }
+    }
+}