+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/utils/extensions/String.kt
@@ -18,7 +18,7 @@ fun String.toMarkdown(): String {
 
 private val markdownV2LinkEscapes = mutableSetOf(')', '\\')
 private val markdownV2PreAndCodeEscapes = mutableSetOf('`', '\\')
-private val markdownV2CommonEscapes = mutableSetOf(
+private val markdownV2CommonEscapes = setOf(
     '_',
     '*',
     '[', ']',