+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/utils/StringFileExtension.kt
@@ -1,0 +1,6 @@
+package com.github.insanusmokrassar.TelegramBotAPI.utils
+
+private val extensionRegex = Regex("[^.]*$")
+
+val String.fileExtension
+    get() = extensionRegex.find(this) ?.value ?: ""