+++ CHANGELOG.md
@@ -42,6 +42,10 @@
 
 * `TelegramBotAPI`:
     * `BotCommand` now will check and throw error in case when command or description lengths is/are incorrect
+    * `StorageFile` now is common for all platforms
+        * JavaScript realization was removed due to its redundancy
+        * JVM realization was replaced with `fun` factory
+        * `StorageFile` now able to accept any factory of `Input`
 
 ## 0.25.0