+++ CHANGELOG.md
@@ -47,7 +47,7 @@
         * JVM realization was replaced with `fun` factory
         * `StorageFile` now able to accept any factory of `Input`
         * `StorageFileInfo` was added to avoid strange collisions with throws in `StorageFile`
-
+    * Fixes issue with `hashTag` for markdown
 ## 0.25.0
 
 * Common: